Add media query for smaller screens
When the screen gets less than 550px
wide, the text in buttons
of the open winner-info-dialog were
wrapping to the next line.

Add a media query for screens smaller
than 500px and reduce the font-size
and paddings of the buttons so that
the button text will not overflow.

On line 104, reduce the padding
left and right of the open
winner-info-dialog and on line 128,
reduce the gap
between its buttons (modal-buttons)

These 2 actions are taken to give the
buttons space so that they do not have
shrink so much to make their text
content to wrap.