Reduce padding of dialog
Remove the padding property and
value of the dialog-container
and allow the open
winner-info-dialog to control the
padding of its contents.

From line 125, give the versus-div
a font-weight of normal. This overrides
the font-weight value given in
dialog-display-scores. A font-weight
of normal makes the player names and
scores to stand out more.