Add form validation
In newGame() on line 515, Assign
the nodeList of all text inputs to
a constant.

Create a checkInputFields() that
first converts the nodeList to an
array so that the some() array
method can be used on it.
Next it uses the some() method
to check if any of the text input
fields are empty. If so, it returns
true, otherwise it returns false.
It then uses the some() method
again to check if any of the
text input fields are greater than
10 characters. If so, it returns
true. Otherwise, it returns false.

It assigns the return value of
two checks to relevant constants.
It then disables the startGameButton
with either of the 2 constants using
the || operator.
Since the || operator returns
the first truthy value, if any
of these checks are false, the
startGameButton will be disabled.

Loop through the nodeList of all
text inputs and add an input event
listener to each one. The input
calls the checkInputFields(),
whenever the user types in any of
the text input fields.

Call checkInputFields() so that
the startGameButton is disabled
when the page loads. Not calling
checkInputFields() will leave the
startGameButton enabled until user
starts typing in a text input.

Limit the length of the text input
to 10 characters because excessively
long names will mess up the structure
of every place in the page that the
name is rendered.

Also update the text inputs placeholder
text on line 477 and 490 because long
placeholder texts seem truncated
on devices with smaller screens
widths.