Add a form-hint
In newGame(), from line 462,
create a new span element that
will be the form-hint. Add a
relevant class and a text content.
On line 504, Append it to
startPageContainer after formDiv.

In the CSS, from line 97 justify it
center (since its parent is a grid
container), give it a color and
a smaller font size.

Also the start-page-container has an
extra declaration on line 103. Move
the height property and value inside
this declaration to line 42, inside
the first declaration.