Start adding color styles
Style the body a dark background color
and on line 216, style the cells'
background with a lighter version of
this color so that the cells standout
from the rest of the page. Style body
and the cells with bright color. After
a winner or a draw, when the modal
opens, all the classes that were
added dependent of marker were removed,
resetting the cells' colors to default.
The default is black and does not look
good on the dark background, thus the
reason for styling the cells with a
lighter color.

Create two custom color properties on
the root selector. One of the colors
is for the x-marker(player one),
while the other is for
o-marker(player two). From line 220,
use these properties to style the color
of the players'labels (on the form)
and their respective markers
(on the board).Use the properties to
also style the players' containers
(in the scores div).