Add more color styles
On line 59, style the backgrounds of
form-h1 and board-h1 with blue gray.
On line 102, style the backgrounds of
the start game and new game buttons
with medium green. From line 153,
style the backgrounds of the the next
round and new round buttons to medium
coral and medium light blue
respectively.

Up in the root selector, make the
x color and o color properties
darker in order to fit in the dark
background.

On line 181, give all buttons a
white color and a bold font weight.