Style the button's states
On lines 106, 167 and 181, style
the buttons' hover and disabled
states. On line 211, update the
color of buttons to whitesmoke,
a darker shade of white. The
hover states of these buttons
all have a color of white.
The transition from whitesmoke
to white, coupled with the darker
shades of these colors (as applied
in the hover states) creates a
good hover effect.