Update scores-div style
Update the scoresDiv in the JS,
such that it is appended to
the gameContainer. Previously
the space between the board
and scores-div was uneven with
the space between the
board-header-div and the board.
Appending  the scoresDiv to
the board allows for a better
styling method to create even
space.

In the CSS, on from line 197,
update the style of
game-container a flex display,
flex-direction of column and
a gap.

On line 295, update the
score-div to justify its contents
with space-evenly. Its former
value - space-between lined it
up with the contents of
board-header-div and it did not
look very pleasant.
Still for the score-div,
on line 305, update the color of
the texts in the players'
containers.

On line 250, increase gap of
board-container to add space
between display-turn-h2 and the
board.