Add styles to the dialog
From line 133, style the dialog
with a background color that is
a lighter shade of the body and
remove its border.

Decrease the font size of the
dialog display-info so that the
dialog-display-scores can have a
larger font size. Style it with
an uppercase text-transform, make
it bold and give it a color that
is a lighter shade of the
form-h1's color.

Increase the font size of the
dialog-display-scores. On
line 238, in the media query,
reduce its font size for
devices with smaller widths
because the large font size
would cause the text to wrap
on these devices (smaller
widths).

Darken the hover
background-color of
next-round-button and
new-round-button for improved
hover effect.

Reduce the gap of
game-container because the
scores seemed to be too close
to the bottom of the page.