Add more styles to dialog
Increase the padding-top
and bottom of the
winner-info-dialog
to create more space at the
top and bottom of the dialog.

Increase the gap of the
dialog-container to create
more space between its
contents.

In the media query, decrease
the font-size of the display-info
and increase the font-size of
the dialog-display-scores. On
devices with smaller widths,
they seemed to be the same
size. The display-info needed
to be smaller than the
dialog-display-scores.

Add display-player-one and
display-player-two on lines
287 and 291 respectively so
that they are styled with the
colors designated for x and
o markers respectively.

Remove the font-weight property
and value of player-one-name,
player-two-name so that the
player names on the scores-div
are not bold. The actual scores
(which are already in bold)
need to be emphasized.