Style the form-div and cells
Style the form div with a
background color that is a
lighter shade of the body
background so that any
box-shadow applied will be
prominent.

Add a box-shadow to the
form-div and cells on line
75.

On 227, remove the border from
button so that all cells can be
without border and so that the
box-shadow effect can be seen
better. Since the border has
been removed from the button
element selector, remove the
border: none from the
individual buttons on lines 101
and 236.

Update the color of form-hint on
line 130. Add padding top
an bottom to create some space
between it and the form-div.
Its parent element is the
start-page-div and the gap could
be used instead of the margin-top
but the space between form-h1 and
form-div, and form-div and form-hint
needs to be uneven.