Add more styles
On line 144, add a semi-transparent
black background to the
This is so that whenever the dialog
is opened, it grabs the attention
of the viewer better.

On line 234 style all active enabled
buttons with transform of
translateY(4px). This moves the
buttons 4px down along the y-axis
(vertical-axis) and gives the buttons
a "pressed" effect.

On line 277, decrease the font-size
of the display-turn-h2 so that it
does not look as big as the board-h1.

Remove the black border from board
as it is no longer needed.