Clean up the javascript
Now that project is almost
deployment-ready, on line 47
update the default cell value
in createCell() and from 0 to
an empty string.

On line 49, in resetCell(), also
update the value the cells are reset
to, from 0 to an empty string.

In placeMarker(), on line 25, since
the default cell value is now an
an empty string, update the condition
to placePlayerMarker, if the cell
value is an empty string.

In handleBoardClicks(), on line 344,
update the condition to
playMove() if the target's event's
textContent is an empty string (no
longer 0).

On line 191, update controlGame()
so that its arguments are the
screenController()'s parameters.
Now calling screenController()
(line 543) inside newGame() can
play the game with the values of
the player names input fields.
It was this way previously but
was changed temporarily (hard coded
arguments in controlGame() on line
191) to allow for easy styling.

Lines 246, 262 are updated
textContent from lower to upper
case markers of the players markers
in the scoresDiv because uppercase
looks better on the scoresDiv.

Lines 371, 277 and 464 are minor
updates to various textContent,
simply for aesthetics.