Clean up HTML
At some point, forgot to add
index.html to staging area and
commit. Wrap every element in
the winner-info-dialog in a
new a dialog-container.

On line 6, chuck a joystick
emoji inside an SVG <text>
element and use that as a
favicon.

Update the title of HTML
document.

In the CSS, increase the font
size of the player scores in the
scores-div so that they stand
out from the markers and the
player names.